+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Next
+# 3.2.0
 
 ## Features: